backend/requirements.txt
aiohttp==3.7.4.post0
python-socketio==5.4.0
